improved dotfinder
The dotfinder now checks if a found dot is circular enough, this will exclude black spots of the cell as these are mostly non circular, this however also tends to exclude dots that are fused together. So we will have to figure out a way to seperate fused together dots.